add compute_tfidf_features utility function
