+++ requirements.txt
@@ -1,6 +1,8 @@
 boto3==1.14.62
+pandas==1.1.3
 regex==2020.10.15
 requests==2.24.0
 sacremoses==0.0.43
+scikit-learn==0.23.2
 sentencepiece==0.1.91
 tqdm==4.50.2
